Get initialization volume from the URL GET params
The URL should end with '?volume=', followed by the desired volume to
initialize (as a percentage. That is, full volume is denoted by
'...?volume=100')

If this is not done, the stream will play at full volume, with the control
set to 20%, because there will be a null error. I will fix this
momentarily.