Carry over set volume through pause/play
This is done by specifying the current value in the play handler (which
refreshes the page). First, the existing params are removed from the URL,
then the volume which is currently set is appended to the URL so that
readyPlayer() will initialize the page with that volume. Then, the page is
reloaded with that URL.