Add themeNames array
I am expanding the concept of a theme, as far as the software is
concerned. A theme will be identified by a theme name and a themeID, which
are interchangable via functions defined in player.js.

This way, the themeID can be stored in the URL just like volume.

The themeID of a theme corresponds to the index of its name in the added
array. The name is a stylized version of that set in the index.