Add fetching and restoring of the theme on load
The theme to initialize can now be set by adding "theme=", followed by a
themeID, to set the theme to the one corresponding to that ID. The
default, if no theme is set in the URL, is chicago-evening.

This is all done in the same way as it was done for volume.