Add validation of the volume and theme set in the URL
Volumes must be between 0 and 100 (negative volumes are nonsensical,
volumes over 100% are erroneous), themes must be a valid index for the
themeNames array (therefore, themeNames must be kept up-to-date).

As a side note, this is the point that I got to yesterday. All commits
before this point that were done by me (at the moment,
18e15ac31722fa8f876f77117f9522d8ed1894a8 is the latest commit this doesn't
apply to) are just me replaying my changes done yesterday into Git (hence
why they are done with erratic timing). This is also why this commit in
particular is signed. All commits after this, however, are newly written.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>