Modify stream loading
The page now only loads the stream while its playing - If it's paused,
then the page will start loading a void, fake OGG URL.

If, of course, the page was just loaded, it will proceed to autoload the
stream. I will fix this..

Additionally, Ken seems to only want this behaviour on mobile. I will
therefore respect his wishes, and implement this behavior only on mobile.