Force video reload when ISS theme is selected - Fixes #24
The last commit caused the browser to no longer load the video, perhaps
because it had 'learned' that the video did not exist. We now reload it
every time the Space Station theme is selected.

This does, however, mean that selecting Space Station causes the video to
restart. I don't want that, so I'll be looking into a fix before I push
this.


Former-commit-id: 5f90153d07feb4c38ffae0d18d47cd9d1dd42bba