Add tokyo-night to theme.json
Thus enabling it to be used via themeChanger directly.

The title is the same as star-guardian and electromaster - 'keidensu',
which is to say, "cadence".

The subtitle is Google Translate's translation of "A Rhythmic Experience"
(chicago-evening's subtitle, and the original subtitle pre-themes) to
Japanese. I can't really tell if this is a good translation, although if
@kenellorando wants to be more serious about this I could ask Mitsuru for
help. Or, someone with a bit more knowledge of Japanese characters could
try to assemble a subtitle which is pronounced close to "A Rhythmic
Experience", although I have my doubts about that being any good.
Basically, the subtitle is almost certainly just temporary.

The theme color is like chicago-evening, but darker - As night is darker
than evening.


Former-commit-id: 45ea882d71ee15cb5376e10fa4f2f63d0acb70ff