Change fonts of tokyo-night title and subtitle
Since Rock Salt and Roboto don't support Japanese characters. I took the
choice of font from electromaster's title, and applied it to tokyo-night.

The italics are good on the title, but I removed it from the subtitle for
being too small (it obscured some detail on the smaller characters of the
subtitle).