Add mobile blocking functionality to themeChanger
The original reason why @kenellorando removed Space Station was that it
was too heavy on mobile, and didn't really look good. Which are valid
problems.

The thing is, on desktop, it's a pretty good theme, and it still isn't
heavy enough that it produces problems.

The original solution was just to kill the theme, because a nice theme on
desktop wasn't worth serious issues on mobile. Which was fair: We had
other things to work on, we were already investing too much work into
making Space Station viable on mobile, and of course the legacy theme
"engine" was absolutely not suited to the kind of work we expect of the
modern engine. So, while I liked the theme, it is indubitable that
removing Space Station was good for Cadence, and a simple diff reveals
just how far we've come.

Far enough, I think, that Space Station can work again. The original
problem still persists: Even trimming the video enough that quality is
noticably worse, Space Station is pretty heavy on mobile. The video
constantly running is just not good for resource limited devices
(lightmage works because the video is so small that it's actually lighter
than chicago-evening even - Minimalist is so far the only theme that is
less heavy than lightmage).

However, the new theme engine presents an opportunity to provide an
extension and have a less extreme solution: When attempting to load the
theme, check if the device is mobile, and if so, load a different theme.
This preserves the high-quality theme on desktop, while still removing the
excessively heavy theme on mobile, and this is the solution I am using to
resurrect Space Station as space-station.

This commit implements that behavior, albeit rigidly - Always loading
chicago-evening. I'll extend it a bit in a little while.


Former-commit-id: 7f56c0d733e5ee38c4577accefe70fc65d8cc2a1