Add earth image
I'll remove this next commit (to comply with Ken's convention of having
theme images hosted on Imgur instead of in the directory).

However, I want the git repository to be able to produce a working version
of Cadence without any outside data (Hence why I worked to fix the LFS
version of the original iss.mp4 on github). So, I'm storing a copy of the
earth.png image here, so if somehow imgur loses the image we still have a
copy.

The image is an edit of a NASA image in the public domain, sourced from
https://apod.nasa.gov/apod/image/0304/bluemarble2k_big.jpg. The image has
been edited to only show Earth, and then letterboxed to make Earth appear
smaller in the image (which is good mostly for mobile).

Former-commit-id: 999f8172be3bb7c18b8a3b5404260c0e8282186c