Readability enhancements
Slightly darken the background, so that it interferes less with the UI.

And, add two text shadows to all text. One is very blurred, making the
entire area near text (especially down and right) appear darker. The other
is closer to the text and more defined, to provide a sharper outline.
These combine to both make the text more legible and to prove something of
a neat effect (as the text appears to cast shadows on the Earth, but not
on space).