Add legacy chicago evening theme
This theme background was the first Cadence Radio theme background. It is
so old, that it predates the name "Cadence Radio", hailing from the time
when Cadence was titled "RADIO" in Permanent Marker.

We've moved forward to Rock Salt and CADENCE, but this image reminds us
all of where we came from.

I pulled this image from the infamous 'F' commit, that is:
7a034694d91657cf87eb671751ca9a4ef381cd8f [formerly e6262038605eee83acd4535c65257d57228fb718]


Former-commit-id: e821ff1e76c5309952eab636a8cda977dd320804