Correctly remember to handle null currentTheme
preLoad is required to handle a value of `null` for currentTheme, which is
generated in the case that the theme being loaded is the default and the
theme is in the process of first being set.

I didn't realize this would be the case when first implementing
LoggingCallback. So here I have to update my example case to actually work
with the code written to run callbacks.