Preliminary support for ARIA patterns
The server will now respond to ARIA POSTs with conformant results.

Neither function makes an attempt to actually interact with the rest of
the Cadence backend. But, both send a reply, and both replies are
"correct". Therefore, this commit fixes #3.

In addition, this commit already supports timeouts for requests (hence the
import for the math library). This is earlier than expected - However,
since @kenellorando hasn't okayed my plan yet, the function doesn't
support enhanced CadenceBot timeout support (where whitelisted IP
addresses may provide a tag to use for timeouts instead of their own
addresses).

This function also extends the power of sendResponse to support having the
invoker provide additional headers to send (as an array of undelimited
strings). That's used extensively in ARIA's new backend, as timeout errors
include a header to specify when the user may submit another request, and
POST requests to locations besides the ARIA API hooks provide a header
informing the browser that the server only requests GET and HEAD methods.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>