Run ARIA requests on a separate thread
This way, the telnet timeout can't block the server.

Eventually, I intend to have all the socket and file operations done in
non-blocking mode. That way, the server won't spend time waiting for them.

However, that does not affect ARIA. Rather than allow telnet (for request)
and the database (eventually; for search) to block the entire server,
these functions will be run on separate threads. The Python interpreter
will still force ARIA to run code in lockstep with the rest of the code
(at least for CPython, the reference interpretation), but the I/O will run
while the interpreter allows the network loop to handle non-ARIA traffic.