Reassure browsers that we are not tracking them
The server does not, and indeed can not, track users, at all, on any URL.
Therefore, we send the Tk header to inform the browser that we're not
tracking the user at all.

In theory, this should be sent in response to DNT headers, but I think
it's harmless to send to all requests.