Enable 418 I'm a teapot easter egg
Requesting a URL ending in "coffee" which would otherwise generate a 404
error results instead in a 418 error, possibly including an image of a
teapot I rendered in Blender with Cycles (if the file can be opened
without trouble).

Gotta have some fun, right?

Also, this is sort of a tradition for me to add when I'm making webservers
that might see public use.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>