Issue redirects for 'accidental' directory paths
Accessing a path which corresponds to a directory via a URL that doesn't
end in either a path separator (for the server's OS) or a slash (/)
results into a redirect to the same URL with a slash at the end.

For example, "cadenceradio.com/original" redirects to
"cadenceradio.com/original/".

This is done with a 301 Permanent Redirect, so browsers can remember to
access the correct path in the future.

This commit closes #20.