Support special request timeouts if configured
This means that whitelisted IP addresses may send requests for songs with
a tag representing a user, and only have the five-minute timeout set for
that user.

This behavior would be pretty nice for CadenceBot.

@kenellorando isn't sure if he wants this feature active. But I've
implemented it here, with a config option that's disabled by default.

If he later decides to flip the switch and turn it on, the switch is
there.