Note that the request blacklist can use tags too
Tags can be found in logs, as they are always logged at level DEBUG, and
logged at level INFO for requests that are blocked by the timeout
mechanism. They should be put in the config file as they are in the logs.

I decided this might be a useful feature in addition to the address
blacklisting. However, both are necessary - an administrator may wish to
block an entire address, or only a single user.