Treat ill-formed requests correctly
Returning a 416 error instead of crashing

Note that as written, this means that multipart requests are disallowed
(they get a 416). If this feature is wanted later, we should add
additional detection before this try.

However, I think that this is fine. Multipart requests are somewhat rare,
especially for Cadence's use case, and denying them should just result in
the browser fetching the parts in multiple requests.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>