Pre-create the database search startswith string
This is a small improvement, but it is one.

I would like to note that this will be my 129th non-merge commit to
cadence-server.

As of this moment, I have 129 non-merge commits merged into the primary
fork of Cadence.

This means that, by commits, this commit marks the point where half my
contributions to the main Cadence program are on the server.