Correct extra log entry on accept
The server was generating a spurious log entry, because it logged before
attempting accept. Therefore, the time accept threw an error would be
logged as the server receiving a connection, even though no such thing was
happening.

As an added bonus, this also logs the incoming socket file descriptor, to
help track socket process.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>