Don't wait for writes to finish before select
The only thing we need to wait for is to remove the connections from
openconn (to avoid double processing). We can do that while we wait for
reads to finish, which lets us go back to responding to connections
massively faster (since writes are almost always bigger than reads for our
purposes).

This is significant for iss.mp4 - On the connection that Testament will be
connected to, the file should take at least three seconds to upload. If
Ken is a building with a slower connection, that would be thirty.

We don't want to wait that long before we accept new connections. This
commit lets us upload that file while also handling other things.