Check for zero length requests earlier
There's not much point in dealing with the request before checking if
it's zero length.

This does carry the problem that the Connection's IP won't be set. I don't
think that matters right now, but it's fair for sendResponse to assume
that IP processing has been done, since this is the only exception. To
that end, I'll initialize the IP on accept, and then see if we can set it
to a better value later.