Init new connections with an IP address
Since the listen socket does provide us with a connecting IP address, we
can use that at first until downloading some headers and seeing what we
get from them.

Also, log the address of the other side of a connection when we connect,
but only at debug level (since it's not yet authoritative data).