Allow config to be loaded from other CWDs
The log code was written to be CWD-independent, but the config code
wasn't, so one still needed to run the server from the directory where it
was stored (doing otherwise would cause the server to crash from inability
to read the config).

Therefore, make sure that we're always trying to access the config files
that are in the same directory as the server code, thus allowing the
server to be run outside that directory.

This is especially useful for a potential run script, which would probably
not usually be run from the server directory.