Add a config option to timeout select calls
I've noticed search responses occasionally not getting sent until after
another connection.

This is, I've realized, because searches are run in another thread. If the
search takes long enough that the response isn't in openconn before the
selection lists are assembled, then the selection ignores the search
response. This means that it has to wait until the next selection to work,
which could be infinite time (until the next connection gets accepted).

Therefore, this option will artifically force recreation of the lists
periodically, so that this behavior is limited.

Currently, this only affects searches and requests, since they're the only
actions that selection doesn't wait to finish that require further action.
Furthermore, it only affects some searches and requests - Those which
occur either alone or with requests that don't outlast them.

However, I intend to add an option to allow reads to behave this way, so
that they don't necessarily block selection for other writes. That will
cause all reads to follow this behavior, so that means that this option is
sorely needed.