Pass timeout to selection
I believe that this should just...work. Both the read code and the write
code is designed to handle having zero sockets, so that won't cause
problems, and we should just go back up to the top.

On the other hand, passing None should trick select into thinking we
didn't give it a timeout at all, reverting to the behavior I documented.

This does cause a /lot/ of logging on debug levels. Debug levels are
intended to have lots of output though, so that's fine.