When maxThreads is 1, don't spawn no-op Threads
When maxThreads!=1, both the read operation and the write operation are
basically no-ops when there's zero sockets for them to work on. This isn't
true when each type of operation is processed on its own thread - These
threads would be spawned, do nothing, then end, incurring the cost of
creating a new thread.

Therefore, skip creating these threads if they're just going to do
nothing.