Load all scripts deferred on non-buggy browsers
IE9 and earlier have a known bug that can break the order of execution of
deferred scripts. On these versions, the behavior I implemented earlier is
needed to ensure that resources used in other scripts are loaded before
executing those scripts.

On browsers which follow standard behavior, that is, which execute
deferred scripts in the order they appear, we can safely defer all the
scripts - While async is technically better, the benefit of deferring
jQuery (the largest script by an order of magnitude) and theme.json (the
second largest script by about 50%) is comparatively large. Thus, I prefer
this behavior on browsers which won't potentially fail to load when it's
available.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>