Protect default config with a special hash
The protection hash is done after stripping away CR and LF characters, so
any line encoding will pass the check.

That is, even though I edit with Windows line endings (CRLF), and
@kenellorando probably has git set up to check the file out with Unix line
endings (LF), the check on his system will pass with the hash made on my
system.

This feature, while critical to the platform-independent deployment of
cadence-server, has one important drawback - Standard tools won't generate
a hash compatible with this check!

Therefore, developers should first run the server with the old (invalid)
canonical hash, and then use the hash printed in the error message to
update the check.

To be specific, the hash chosen is SHA256, with the aforementioned
characters stripped. The choice of stripped characters is specific to the
solution of the above problem, allowing @kenellorando's *nix machine and
my Windows machine to pass the same check. The choice of hash is for
consistency with the ETag - I'd rather not unnecessarily duplicate the
hash algorithms used by the server code.

This commit fixes #40.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>