Execute theme and player setup earlier
These are executed on document ready instead of on body load. This makes
the load feel snappier, as the page gets themed and the volume gets set
before nowplaying information and the music stream start loading.

This is especially noticable in development - If these two don't actually
start loading, the theme never gets set, but with this commit, it loads
almost instantaneously.