Remove Node.js remnants
This commit completes the task of integrating the new server code into the
mainline cadence repository. As such, it affects a number of items:

 - Closes #117 (invalidated due to removal of Node server)
 - Closes #118 (invalidated due to removal of Node server)
 - Fixes #101 (new server implements special search forms - "<genre>
   songs" is effectively a search-by-genre
 - Fixes #98 (New server does not use any CORS headers in default
   configuration)
 - Closes #69 (invalidated due to removal of Node server)

The server code is not currently complete. For one, requests appear to not
yet be working, and we'll be working on that. Issues remaining open on the
server are:

 - Implement keepalive connections
 - Consider reducing conversions to and from unicode
 - Break up server.py into multiple files (so each file handles one part
   of the server code)
 - Implement proper handling of special characters (Assigned to
   @Karenwithakay)
 - Support HTTP/2.0
 - Minify HTML/CSS/JS files
 - Remove block between disk and network (I really don't care enough to
   fix this)
 - Remove block during disk I/O (See above; also, as far as I can tell
   that will break Windows support)

Additionally, I would like to add a script to perform server startup
without having to type out the command.

However, this commit is the one where the Python server takes over for the
Node server, so this is the one that does all of these things.

And here, I will say: It has been my honor to develop and have tenure over
the new Cadence server. I hope that I have accomplished the goals in the
project: To create a server which is more stable than the legacy Node one;
More feature-rich than the legacy one; And one which we are better able to
maintain and develop features for than the legacy one.

I am grateful to have been entrusted with this project, and I hope for
great things with respect to it's future under the tenure of
@kenellorando.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>