Add documentation for randomized theme behavior
This feature was recommended by @JakobFrank.

Basically, if the preLoad handler returns true (the boolean), we take it
as a signal to load any theme we want, which in this case will be random
(since if the intent was to not change, it would return the passed
currentTheme).

Since this was his idea, I'm marking him as a co-author - Thus, with this
commit, he becomes the fifth person to receive credit for work with
Cadence. Congratulations, @JakobFrank!

Co-authored-by: Jakob Frank <31579263+JakobFrank@users.noreply.github.com>
Signed-off-by: Ryan Hodin <ryan951357@gmail.com>