Add theme object for random theme
This is a theme which copies chicago-evening, even claiming to be
chicago-evening, except when it's loaded, it instructs the engine to
choose another theme.

Unfortunately, even if it didn't emulate chicago-evening in every way,
there wouldn't be a way to load a random theme on load - The engine
simply doesn't support that (since setting the random theme also sets the
default theme).

Therefore, in case the callback interface is broken, random falls back to
chicago-evening.