Add themeChoice button for random theme
Placing the theme at the bottom of the rotation, as I think that it's
interesting and novel.

I'd sort of like to be able to set the defaultTheme to load a random theme
- that would be cool.

But current architecture doesn't really support that, since the default
theme is set on theme change, and random can never be changed to (since
changing to it will immediately change to another theme).

The easiest way to support this would be to add a switch in localStorage,
storing whether we should use the actual default theme or random. This
switch would then be set whenever the themeChanger is called on a themeKey
"random", and unset whenever it's changed to another theme.

This could either be placed in the themeChoice onClick handler, or it
could go in the themeChanger function itself, as long as it was checked
prior to the preLoad handling loop (since the function isn't recursive for
preLoad changes).

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>