Add nebula_rotating video file
This is an h.264 encode of a crop of the original FFV1 master (which is
itself an encoding of a PNG image series), intended to serve as a file for
widespread distribution as the background for a video theme
'nebulaRotating'.

This file is about twice the size of iss.mp4, which has been the biggest
file in cadence for a very long time. Therefore, it strictly requires the
creation of a mobile version which is smaller and lower quality, which I
shall create later on.

The original source material for this video is a Blender render of a full
rotation of the background image for the nebula theme.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>