Update nebula_rotating.mp4
The old copy had full chroma sampling. That is, the pixel format was
yuv4:4:4. Some browsers apparently don't like that, so I've reencoded this
to be yuv4:2:0, which is more widely supported.

As part of this, @JakobFrank and I discovered that the chroma subsampling
was causing the washout that led me to use FFmpeg directly instead of
handbrake to begin with. So together, we adjusted the gamma of five-second
samples until we found a result we liked, which is the setting used here.

It's slightly darker than the original render, but we agreed that it's
better this way.

For posterity, this encode was performed with the following FFmpeg
commandline:

ffmpeg -i /f/storage/master.mkv -c:v libx264 -crf 18 -movflags +faststart
-preset veryslow -filter "crop=1920:1080:1016:562, eq=gamma=0.8" -pix_fmt
yuv420p nebula_rotating.mp4

Where /f/storage/master.mkv is my FFV1 master for the nebula_rotating
video.

Co-authored-by: Jakob Frank <31579263+JakobFrank@users.noreply.github.com>
Signed-off-by: Ryan Hodin <ryan951357@gmail.com>