Add nebula-rotating themeChoice button
In rotation, since the standard nebula is not.

While this theme is different from the engine's point of view, and
therefore will be saved differently if the user selects static or rotating
nebula, I see no reason why it must be different from the user's point of
view. Therefore, it has the same name as the static nebula themeChoice
button.

Thanks to her contribution in conceptualizing this theme, @Karenwithakay
is a coauthor on this commit.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>
Co-authored-by: Karen Santos <kjaysantos@gmail.com>