Add MIME type detection for Matroska videos
Using the recommended type found at
https://matroska.org/technical/specs/notes.html

This commit is needed for my current Cadence frontend work - I'm adding
support for serving video to users encoded in HEVC, which I'm containing
in a Matroska file. Therefore, this commit is needed for that feature to
work properly (on those browsers which support HEVC, of course).

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>