First-pass shortscreen fix
This fix is rather invasive of the layout.

The way it works is, when the layout is short enough to trigger the
overlapping bug (tested experimentally, on my instance of Chrome on
Windows 10, as between 500 and 550 pixels), the layout switches from being
laid out as flowing elements, to a more fixed layout.

Essentially, the main element, containing the radio player and controls,
is removed from the layout entirely, by being positioned absolutely
(because there's no relative positioned element before it, this means that
it's removed from the layout for the rest of the elements and it's
positioned relative to the body), with width matching the body, and
top-left corner aligned to the top-left corner of the body.

Then, the #aria element, containing all functionality related to song
requests, is given a top margin of 550 pixels - The tallest viewport that
causes the bug, and therefore tall enough that the overlap cannot occur.

As far as the browser is concerned, the main element is laid over aria's
top margin.

It's functionally a lot like the existing layout (I was originally testing
with this enabled at all times, then made the media query for it to only
trigger when necessary), but it does have a few minor differences, and
therefore I'd like to avoid this trick if I can find a way.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>