Switch all selector unregisters to exception-safe
In the live server logs for Cadence, we have a critical-level log entry
(reserved for an uncaught exception which results in a full networking
restart) caused by a double unregister on a socket with file descriptor 8
- Which shouldn't be possible, as far as I can tell. Which means I can't
  really fix it.

However, a full networking restart is bad. At minimum, it means we lose
recursion depth and stutter our service (as it only really drops for a few
milliseconds) - However, it could well result in an inconsistent state,
depending on how it occurs. I want to, as a matter of principle, eliminate
any sources of these errors.

Therefore, when a double-unregister (or any other unregister exception)
occurs, it will be logged at error-level, but control will be returned
normally, allowing remaining operations to occur and keeping Cadence
running smoothly.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>