Import search syntax from v3 server
These searches have been directly ported from Python, and I'm not
entirely sure if I did so correctly, so some review and testing is
necessary.

Once this is fixed up, the general search we have now will be improved
to return results in order of how similar they are to the search term.

Additionally, there will be a number of ways to express that you are
searching for a specific field in natural-ish ways: For example, you
could search for "rock songs", or "songs released in 2015", or "songs
by fripside". Results for these will be more specific because each will
use only one field - Two of which aren't used in the general search -
and they will also be sorted by similarity to the search term.