Revert stream latency update
The stream patch works really well for non-mobile devices except if it is the *first* time playing the stream after page load. With the new patch, the first play will begin playback from the time the _page_ itself loaded. If you don't press play for long enough, you'll end up with a super delayed stream. This is just patching it for now.